move dextester to dex adapter
