+++ modules/contracts/adapters/dex/Cargo.toml
@@ -32,7 +32,7 @@ interface = [
   "abstract-dex-standard/interface",
 ]
 schema = ["abstract-adapter/schema"]
-testing = ["abstract-wyndex-adapter/local", "abstract-astroport-adapter/local"]
+testing = ["dep:abstract-client", "dep:cw-orch", "dep:cw20"]
 
 # Supported Dexes
 wynd = [
@@ -119,6 +119,9 @@ terraswap = { version = "2.8.0", optional = true }
 
 # Astrovault #
 abstract-astrovault-adapter = { workspace = true }
+
+# Testing
+abstract-client = { workspace = true, optional = true }
 
 [dev-dependencies]
 abstract-interface = { workspace = true, features = ["daemon"] }
