+++ modules/contracts/adapters/dex/src/lib.rs
@@ -17,6 +17,9 @@ pub mod host_exchange {
     pub use abstract_osmosis_adapter::dex::Osmosis;
 }
 
+#[cfg(feature = "testing")]
+pub mod dex_tester;
+
 #[cfg(feature = "interface")]
 pub mod interface {
     use crate::{contract::DEX_ADAPTER, msg::*, DEX_ADAPTER_ID};
@@ -28,19 +31,18 @@ pub mod interface {
     use abstract_dex_standard::raw_action::DexRawAction;
     use abstract_interface::{AbstractAccount, AbstractInterfaceError};
     use abstract_interface::{AdapterDeployer, RegisteredModule};
-    use abstract_sdk::base::Handler;
     use abstract_sdk::features::ModuleIdentification;
     use cosmwasm_std::{Decimal, Empty};
     use cw_asset::{AssetBase, AssetInfoBase};
     use cw_orch::{build::BuildPostfix, interface};
     use cw_orch::{contract::Contract, prelude::*};
-
+    
     #[interface(InstantiateMsg, ExecuteMsg, QueryMsg, Empty)]
     pub struct DexAdapter<Chain>;
-
+    
     // Implement deployer trait
     impl<Chain: CwEnv> AdapterDeployer<Chain, DexInstantiateMsg> for DexAdapter<Chain> {}
-
+    
     impl<Chain: CwEnv> Uploadable for DexAdapter<Chain> {
         fn wrapper(&self) -> <Mock as TxHandler>::ContractSource {
             Box::new(ContractWrapper::new_with_empty(
@@ -51,14 +53,14 @@ pub mod interface {
         }
         fn wasm(&self) -> WasmPath {
             artifacts_dir_from_workspace!()
-                .find_wasm_path_with_build_postfix(
-                    "abstract_dex_adapter",
-                    BuildPostfix::<Chain>::ChainName(self.get_chain()),
-                )
-                .unwrap()
+            .find_wasm_path_with_build_postfix(
+                "abstract_dex_adapter",
+                BuildPostfix::<Chain>::ChainName(self.get_chain()),
+            )
+            .unwrap()
         }
     }
-
+    
     impl<Chain: CwEnv> DexAdapter<Chain> {
         /// Swap using ans resolved assets
         pub fn ans_swap(
@@ -70,7 +72,7 @@ pub mod interface {
         ) -> Result<(), AbstractInterfaceError> {
             let asset = AssetEntry::new(offer_asset.0);
             let ask_asset = AssetEntry::new(ask_asset);
-
+            
             let swap_msg = crate::msg::ExecuteMsg::Module(adapter::AdapterRequestMsg {
                 proxy_address: None,
                 request: DexExecuteMsg::AnsAction {
@@ -111,15 +113,15 @@ pub mod interface {
                 },
             });
             account
-                .manager
-                .execute_on_module(DEX_ADAPTER_ID, swap_msg)?;
+            .manager
+            .execute_on_module(DEX_ADAPTER_ID, swap_msg)?;
             Ok(())
         }
     }
 
     impl<Chain: CwEnv> RegisteredModule for DexAdapter<Chain> {
         type InitMsg = Empty;
-
+        
         fn module_id<'a>() -> &'a str {
             DEX_ADAPTER.module_id()
         }
@@ -128,25 +130,26 @@ pub mod interface {
             DEX_ADAPTER.version()
         }
     }
-
+    
     impl<Chain: CwEnv> From<Contract<Chain>> for DexAdapter<Chain> {
         fn from(contract: Contract<Chain>) -> Self {
             Self(contract)
         }
     }
-
+    
     impl<Chain: cw_orch::environment::CwEnv> abstract_interface::DependencyCreation
-        for DexAdapter<Chain>
+    for DexAdapter<Chain>
     {
         type DependenciesConfig = cosmwasm_std::Empty;
-
+        
         fn dependency_install_configs(
             _configuration: Self::DependenciesConfig,
         ) -> Result<
-            Vec<abstract_core::manager::ModuleInstallConfig>,
-            abstract_interface::AbstractInterfaceError,
+        Vec<abstract_core::manager::ModuleInstallConfig>,
+        abstract_interface::AbstractInterfaceError,
         > {
             Ok(vec![])
         }
     }
 }
+
